[#5 2.3.2 Degrees and Radians] Add degree-radian conversion functions

+++ src/Utilities/Constants.h
@@ -4,14 +4,14 @@
 #include <stdlib.h>
 #include "RGBColor.h"
 
-const double 	PI 			= 3.1415926535897932384;
-const double 	TWO_PI 		= 6.2831853071795864769;
-const double 	PI_ON_180 	= 0.0174532925199432957;
-const double 	invPI 		= 0.3183098861837906715;
-const double 	invTWO_PI 	= 0.1591549430918953358;
+const float 	PI 			= 3.1415926535897932384;
+const float 	TWO_PI 		= 6.2831853071795864769;
+const float 	PI_ON_180 	= 0.0174532925199432957;
+const float 	invPI 		= 0.3183098861837906715;
+const float 	invTWO_PI 	= 0.1591549430918953358;
 
-const double 	kEpsilon 	= 0.0001; 
-const double	kHugeValue	= 1.0E10;
+const float 	kEpsilon 	= 0.0001; 
+const float		kHugeValue	= 1.0E10;
 
 // TODO(fix) move colors to RGBColor
 const RGBColor black(0.0);
